backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
camel-tools
passlib[bcrypt]
python-jose[cryptography]
numpy
pandas
torch
faiss-cpu
transformers
requests
sqlalchemy
psycopg2-binary
pydantic
openpyxl
transformers
sentencepiece